Enhance project structure and coding guidelines
Updated README.md to reference CODING_GUIDELINES.md and expanded contributing section. Refactored ResultExtensions.cs for better error handling. Removed ServiceCollectionExtensions.cs and integrated its functionality into DependencyInjection.cs, which now includes IJwtOptionsAccessor for JWT configuration. Updated command classes to improve configurability and logging practices. Added comprehensive CODING_GUIDELINES.md for coding standards and best practices.
