+++ src/Rpssl.Api/Extensions/ServiceCollectionExtension.cs
@@ -19,7 +19,7 @@ internal static class ServiceCollectionExtension
             })
             .AddApiExplorer(options =>
             {
-                options.GroupNameFormat = "'v'VVV"; // v1, v1.1, v2
+                options.GroupNameFormat = "'v'VVV";
                 options.SubstituteApiVersionInUrl = true;
             });
         return services;
