+++ src/Rpssl.Application/Users/Login/LoginUserCommandHandler.cs
@@ -1,5 +1,4 @@
 using Rpssl.Application.Abstractions;
-using Rpssl.Application.Token;
 using Rpssl.Domain.RefreshTokens;
 using Rpssl.Domain.UnitOfWork;
 using Rpssl.Domain.Users;
@@ -12,6 +11,7 @@ internal sealed class LoginUserCommandHandler(
     IRefreshTokenRepository refreshTokens,
     IPasswordHasher hasher,
     IJwtTokenService jwt,
+    IJwtOptionsAccessor jwtOptions,
     IDateTimeProvider dateTime,
     IUnitOfWork uow) : IRequestHandler<LoginUserCommand, Result<UserDto>>
 {
@@ -32,7 +32,7 @@ internal sealed class LoginUserCommandHandler(
         string rawRefresh = jwt.GenerateRefreshTokenRaw();
         string refreshHash = jwt.HashRefreshToken(rawRefresh);
         DateTimeOffset now = dateTime.UtcNow;
-        DateTimeOffset refreshExpires = now.AddDays(14);
+        DateTimeOffset refreshExpires = now.AddDays(jwtOptions.RefreshTokenDays);
         RefreshToken rt = new(user.Id, refreshHash, sessionId, now, refreshExpires);
         await refreshTokens.AddAsync(rt, cancellationToken);
         await uow.SaveChangesAsync(cancellationToken);
