+++ src/Rpssl.Infrastructure/Authentication/JwtOptionsAccessor.cs
@@ -1,0 +1,9 @@
+using Microsoft.Extensions.Options;
+using Rpssl.Application.Abstractions;
+
+namespace Rpssl.Infrastructure.Authentication;
+
+internal sealed class JwtOptionsAccessor(IOptions<JwtOptions> options) : IJwtOptionsAccessor
+{
+    public int RefreshTokenDays => options.Value.RefreshTokenDays;
+}
